Refactor ContributionsAggregate service into separate service
